refactor(l5): move tiering from src/ into server/domain behind an L5 adapter (issue 864)
Moved:
- src/tiering.ts -> server/domain/tiering.ts

The server/ version meets the server/ standard whole: two exported functions
were over the four-parameter rule, so their trailing positional arguments fold
into one options object each.

  graduateLaneEvent(pool, event, namespace, opts: GraduateLaneEventOptions)
    opts: { createdBy, embedding, reason, embedFn? }
  findDurableDuplicate(pool, namespace, lookup: DuplicateLookup)
    lookup: { hash, embedding, threshold? }

Every other export keeps its signature. Behavior is unchanged.

src/tiering.ts is now the L5 adapter (M9): it re-exports the server/ module and
keeps the pre-864 positional call form of both functions behind a labelled rest
tuple, so scripts/tier-lane-durable.ts, scripts/done-means/605-*.driver.ts and
src/tools/tier-lane.ts are untouched. Its every relative import names a server/
path, types included.

server/tools/tier-lane.ts is the runtime importer and now reads
../domain/tiering.ts; server/domain/candidate-dedupe.test.ts reads ./tiering.ts.
Imports of modules still in src/ (embedding, logger, observability, chunk-write)
stay relative per M3.

python/openbrain-memory/tests/test_event_vocabulary.py reads the EventType union
as text; both the read at test_tiering_union_matches_python and the allowed-files
list now name server/domain/tiering.ts.

src/tiering.test.ts stays at its src/ path per M1 EXCEPTION and passes unchanged
through the adapter.

Runtime closure of src/ under server/main.ts: 30 -> 29.

Verified:
- bunx tsc --noEmit -> exit 0
- oxlint --deny-warnings on every staged path -> exit 0
- bun run test:isolated (tiering + candidate-dedupe + tier-lane suites) -> 162 pass, 0 fail
- bash scripts/done-means/864-moved-out-of-src.sh -> judged=36 shims, 2 adapters PASS
- uv run pytest -q tests/test_event_vocabulary.py -> 9 passed; uv run ruff check src tests -> All checks passed
